requirements.txt
fastmcp>=2.0.0
httpx[http2]>=0.27.0
python-chess>=1.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
uvicorn>=0.30.0
starlette>=0.37.0
